Remove the destination file before syncing
Refactor the password syncing code to remove some duplication and make
it clear which command is going to run.  Then, before running the sync
command, remove the destination file (if it exists) so that we don't
get interactive prompts from cp(1) asking if we want to overwrite the
destination file.